more linting

+++ csrc/multidevice/utils.cpp
@@ -33,13 +33,13 @@ bool isSharded(TensorView* tv) {
   return is_sharded;
 }
 
-int dimWithParallelType(TensorView* tv, ParallelType pt, bool withReductions) {
+int64_t dimWithParallelType(TensorView* tv, ParallelType pt, bool withReductions) {
   auto ids = withReductions
       ? tv->getMaybeRFactorDomain()
       : TensorDomain::noReductions(tv->getMaybeRFactorDomain());
   for (size_t i = 0; i < ids.size(); ++i) {
     if (ids[i]->getParallelType() == pt) {
-      return i;
+      return static_cast<int64_t>(i);
     }
   }
   return -1;
